fix dispatcher usage (avoid havoc'ing when calling system contracts)
also add a new invariant and a WIP rule + definitions, and tune the run config

managed to get a run that has both `operatorIdIsUnique` and `registeredOperatorsHaveNonzeroBitmaps` passing! link here:
https://prover.certora.com/output/83341/1879f737069c4467a1671dba4fbc007c?anonymousKey=0c41dccd5f3fc10265595c5628e685326e09bc1d

this is currently taking a stupid amount of time to run (~4k secs / 1.25 hours per invariant). will need to look into ways to trim that down.