preliminary undelegation simplification
This commit moves much of the undelegation functionality and related checks out of the StrategyManager & EigenPodManager contracts, and into the DelegationManager contract itself. The change is intended to better differentiate the core contracts / isolate state / increase the checking of conditions "in the proper place" / reduce calling back-and-forth between contracts.
I've labelled the commit as "preliminary" since it breaks a number of tests, contains several related "TODO" comments, and could still use cleanup in terms of in-code comments.