feat: beacon chain full and half slashes (#1163)
**Motivation:**

We need to test full slashes for EigenPods.

**Modifications:**

Added the `SlashType` enum which slashes either the normal amount (10
gwei), half the validators balance, or fully slashes the validator. This
needs to be passed in on every call to `slashValidators`

Tests:
- [x] test_fullSlash_Delegate
- [x] test_fullSlash_Revert_Redeposit

**Result:**

Expressive BC slashing.

Note, delegation does not revert when the slashing factor of a staker is
zero.