docs: Explaining Minimum Slashable Stake in Better Detail (#1225)
**Motivation:**

Per and AVS builder question:
>For the slashing release, which function calls can the AVS
(servicemanager) invoke to get a given Operator's currently allocated
(slashable) stake amounts for a given OperatorSet?
The answer is not obvious from our current documentation and would
require them to dig through our codebase to interpret the correct
answer. However, this is a very important topic for AVS builders.

**Modifications:**

I've suggested the highest level possible explanation for how to get
this information and where to find more information.

**Result:**

Easier opportunities for developers to self serve with this important
topic.
Additional lines of markdown docs for allocationmanager.md.