fix: taskMailbox audit fixes (#1604)
**Motivation:**

Certora TaskMailbox audit fixes for Hourglass Part 1 and Part 2.

**Modifications:**

* `H-03: Aggregator TOCTOU Issues Regarding Stake Weights and Operator
Set` : Added a `MAX_TASK_SLA` immutable that will be set as
`DEALLOCATION_DELAY / 2` so that AVSs have half the Deallocation Delay
to do any operator slashing in case of misbehavior

* `L-01: TaskMailbox::createTask() may create tasks that cannot be
completed` : Checking that `block.timestamp + taskConfig.taskSLA <=
operatorTableReferenceTimestamp + maxStaleness` during taskCreation so
that a task cannot be created if its max response time breaches the
staleness period of the certificate.

* `L-03: Restrictive check in_validateBN254Certificate()` : Updated the
check to only be for the (0,0) coordinate.

* `I-02: Incorrect NatSpec in registerExecutorOperatorSet()` : Clearer
natspec.

* Updated release scripts

* Updated Bindings

* Updated Docs

* Updated Unit tests

**Result:**

Bug free code.