feat: electra timing fix (#1620)
# v1.6.1 Electra Timing Fix

Fixes a bug on EigenPods regarding partial withdrawals for Electra.
Please read the
[explainer](https://hackmd.io/@rNlehaDSRBydbF_i_McDcw/ByyodLJ9ex) for a
detailed description of the bug. **No action is needed and no customer
funds are at risk.**

## Release Manager

@ypatil12 @nadir-akhtar @gpsanant 

## Highlights

🐛 *Bug Fixes*
- Update the `EigenPod.requestWithdrawal` function to ensure that
validators are pointed to the pod, matching the behavior of
`requestConsolidation`

🔧 *Improvements*
- Update the `EigenPod.verifyWithdrawalCredentials` function to only
accept `beaconTimestamps` that are after the
`latestCheckpointTimestamp`. This enables the eigenpod state machine to
be easier to be reasoned about

**Full Changelog**:
https://github.com/Layr-Labs/eigenlayer-contracts/compare/v1.6.0...1.6.1